execution/abi/bind, db/kv/stream: make test files external
